Upgraded dependencies
Previously there were 12 security issues with npm modules that were
dependencies of the wallet app. After this commit, there will be only
five security issues. This commit upgrades most of the npm packages
to new versions. An effort will still need to be made to upgrade
submodules dependencies to make sure the app is free of any security
issues.